feat: Add month selector to transaction page to filter results
